main.py: Shut off the motors ASAP whenever there is an exception, and use 'Exception' instead of 'BaseException' so that code will not apply to SystemExit or KeyboardInterrupt.
It might look wasteful to import the entire library at the top of
this file, but we currently do not actually have a way of importing
just a portion of the library: __init__.py always runs whenever we
import any piece of hte library.

Also, splash_loader imports the entire library on one of its first
lines, before displaying the splash screen.

Importing the library beforehand gives us me more confidence that
we can shut down the motors and do it quickly.