MicroPython: Improve QTR reading consistency
I noticed an issue with inconsistencies in the QTR readings when
porting the PIO and corresponding QTR code from MicroPython to my
3pi+ 2040 Arduino Library project:
  https://github.com/adamgreen/pololu-3pi-plus-2040-arduino-library

Increasing the charge time to 200us like in the MicroPython code worked
around the problem but originally I just ported the CPU version of the
QTR reading code from the Pololu 3pi+ 32U4 Arduino Library and it only
needed 10us of charge time to get consistent readings. This made me
suspicious that something about the PIO code path that I ported from
MicroPython contributed to the inconsistent readings I was seeing with
shorter charge times. After attaching an oscilloscope to one of the
QTR sensor pins I saw what was going on and was able to track down the
issue in the PIO code path:
* Originally the QTR driver would have the PIO code start the charging
  phase at the end of the previous read.
* The CPU was responsible for kicking off the next read by the PIO.
* The PIO would then start the discharge phase and start timing the
  discharge for each light sensor pin.
* Depending on what the CPU was used to do after the previous read
  completed and before starting the next read, the capacitors may have
  ended up with a different level of charge and this would impact the
  time it took to discharge.
* The PIO code should be responsible for the timing of the charging
  phase as well as the discharging phase or the readings can be
  inconsistent.

This commit includes a change which makes the PIO code responsible for
charging the capacitor for 8us, consistently on each and every sample,
before it switches the sensor pins to be inputs and measuring the
discharge time. I also removed the CPU delay as it is now taken care of
by the PIO code itself.