Better Python exception handling.
If an experimental MicroPython patch is enabled, show the line numbers
of the exception's stack trace on the OLED.  This part might change
later.

Also, store the exception as a local variable so it can be
easily accessed in the REPL.

Also import sys so it's easier to see these exceptions in the REPL.

Note: If the code for displaying the exception goes awry, we can
access that exception with sys.exc_info()[1], if it's enabled.