Reduce QTR PIO program size
I moved the initialization of the OSR, Y, and X registers from the
start of the PIO program and put it into the QTRSensors.run() method
which already existed to reset the program counter back to the beginning
of the program to start each sensor read. This reduced the size of the
PIO code by 3 instructions.

NOTE: This change requires a patch to MicroPython which extends the
rp2.StateMachine.exec() method to accept pre-encoded machine code that
can be found in my
https://github.com/adamgreen/micropython/commits/rp2_state_machine_exec
branch.